Module aws_sdk_lakeformation::error

Modules

access_denied_exception

See AccessDeniedException

already_exists_exception

See AlreadyExistsException

concurrent_modification_exception

See ConcurrentModificationException

entity_not_found_exception

See EntityNotFoundException

glue_encryption_exception

See GlueEncryptionException

internal_service_exception

See InternalServiceException

invalid_input_exception

See InvalidInputException

operation_timeout_exception

See OperationTimeoutException

resource_number_limit_exceeded_exception

See ResourceNumberLimitExceededException

Structs

AccessDeniedException

Access to a resource was denied.

AddLFTagsToResourceError

AlreadyExistsException

A resource to be created or added already exists.

BatchGrantPermissionsError

BatchRevokePermissionsError

ConcurrentModificationException

Two processes are trying to modify a resource simultaneously.

CreateLFTagError

DeleteLFTagError

DeregisterResourceError

DescribeResourceError

EntityNotFoundException

A specified entity does not exist

GetDataLakeSettingsError

GetEffectivePermissionsForPathError

GetLFTagError

GetResourceLFTagsError

GlueEncryptionException

An encryption operation failed.

GrantPermissionsError

InternalServiceException

An internal service error occurred.

InvalidInputException

The input provided was not valid.

ListLFTagsError

ListPermissionsError

ListResourcesError

OperationTimeoutException

The operation timed out.

PutDataLakeSettingsError

RegisterResourceError

RemoveLFTagsFromResourceError

ResourceNumberLimitExceededException

A resource numerical limit was exceeded.

RevokePermissionsError

SearchDatabasesByLFTagsError

SearchTablesByLFTagsError

UpdateLFTagError

UpdateResourceError

Enums

AddLFTagsToResourceErrorKind

BatchGrantPermissionsErrorKind

BatchRevokePermissionsErrorKind

CreateLFTagErrorKind

DeleteLFTagErrorKind

DeregisterResourceErrorKind

DescribeResourceErrorKind

GetDataLakeSettingsErrorKind

GetEffectivePermissionsForPathErrorKind

GetLFTagErrorKind

GetResourceLFTagsErrorKind

GrantPermissionsErrorKind

ListLFTagsErrorKind

ListPermissionsErrorKind

ListResourcesErrorKind

PutDataLakeSettingsErrorKind

RegisterResourceErrorKind

RemoveLFTagsFromResourceErrorKind

RevokePermissionsErrorKind

SearchDatabasesByLFTagsErrorKind

SearchTablesByLFTagsErrorKind

UpdateLFTagErrorKind

UpdateResourceErrorKind