Module aws_sdk_lakeformation::model

Modules

batch_permissions_failure_entry

See BatchPermissionsFailureEntry

batch_permissions_request_entry

See BatchPermissionsRequestEntry

catalog_resource

See CatalogResource

column_lf_tag

See ColumnLfTag

column_wildcard

See ColumnWildcard

data_lake_principal

See DataLakePrincipal

data_lake_settings

See DataLakeSettings

data_location_resource

See DataLocationResource

database_resource

See DatabaseResource

details_map

See DetailsMap

error_detail

See ErrorDetail

filter_condition

See FilterCondition

lf_tag

See LfTag

lf_tag_error

See LfTagError

lf_tag_key_resource

See LfTagKeyResource

lf_tag_pair

See LfTagPair

lf_tag_policy_resource

See LfTagPolicyResource

principal_permissions

See PrincipalPermissions

principal_resource_permissions

See PrincipalResourcePermissions

resource

See Resource

resource_info

See ResourceInfo

table_resource

See TableResource

table_wildcard

See TableWildcard

table_with_columns_resource

See TableWithColumnsResource

tagged_database

See TaggedDatabase

tagged_table

See TaggedTable

Structs

BatchPermissionsFailureEntry

A list of failures when performing a batch grant or batch revoke operation.

BatchPermissionsRequestEntry

A permission to a resource granted by batch operation to the principal.

CatalogResource

A structure for the catalog object.

ColumnLfTag

A structure containing the name of a column resource and the tags attached to it.

ColumnWildcard

A wildcard object, consisting of an optional list of excluded column names or indexes.

DataLakePrincipal

The AWS Lake Formation principal. Supported principals are IAM users or IAM roles.

DataLakeSettings

A structure representing a list of AWS Lake Formation principals designated as data lake administrators and lists of principal permission entries for default create database and default create table permissions.

DataLocationResource

A structure for a data location object where permissions are granted or revoked.

DatabaseResource

A structure for the database object.

DetailsMap

A structure containing the additional details to be returned in the AdditionalDetails attribute of PrincipalResourcePermissions.

ErrorDetail

Contains details about an error.

FilterCondition

This structure describes the filtering of columns in a table based on a filter condition.

LfTag

A structure that allows an admin to grant user permissions on certain conditions. For example, granting a role access to all columns not tagged 'PII' of tables tagged 'Prod'.

LfTagError

A structure containing an error related to a TagResource or UnTagResource operation.

LfTagKeyResource

A structure containing a tag key and values for a resource.

LfTagPair

A structure containing a tag key-value pair.

LfTagPolicyResource

A structure containing a list of tag conditions that apply to a resource's tag policy.

PrincipalPermissions

Permissions granted to a principal.

PrincipalResourcePermissions

The permissions granted or revoked on a resource.

Resource

A structure for the resource.

ResourceInfo

A structure containing information about an AWS Lake Formation resource.

TableResource

A structure for the table object. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

TableWildcard

A wildcard object representing every table under a database.

TableWithColumnsResource

A structure for a table with columns object. This object is only used when granting a SELECT permission.

TaggedDatabase

A structure describing a database resource with tags.

TaggedTable

A structure describing a table resource with tags.

Enums

ComparisonOperator

DataLakeResourceType

FieldNameString

Permission

ResourceShareType

ResourceType