Module aws_sdk_lakeformation::operation

Structs

AddLFTagsToResource

Attaches one or more tags to an existing resource.

BatchGrantPermissions

Batch operation to grant permissions to the principal.

BatchRevokePermissions

Batch operation to revoke permissions from the principal.

CreateLFTag

Creates a tag with the specified name and values.

DeleteLFTag

Deletes the specified tag key name. If the attribute key does not exist or the tag does not exist, then the operation will not do anything. If the attribute key exists, then the operation checks if any resources are tagged with this attribute key, if yes, the API throws a 400 Exception with the message "Delete not allowed" as the tag key is still attached with resources. You can consider untagging resources with this tag key.

DeregisterResource

Deregisters the resource as managed by the Data Catalog.

DescribeResource

Retrieves the current data access role for the given resource registered in AWS Lake Formation.

GetDataLakeSettings

Retrieves the list of the data lake administrators of a Lake Formation-managed data lake.

GetEffectivePermissionsForPath

Returns the Lake Formation permissions for a specified table or database resource located

GetLFTag

Returns a tag definition.

GetResourceLFTags

Returns the tags applied to a resource.

GrantPermissions

Grants permissions to the principal to access metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3.

ListLFTags

Lists tags that the requester has permission to view.

ListPermissions

Returns a list of the principal permissions on the resource, filtered by the permissions of the caller. For example, if you are granted an ALTER permission, you are able to see only the principal permissions for ALTER.

ListResources

Lists the resources registered to be managed by the Data Catalog.

PutDataLakeSettings

Sets the list of data lake administrators who have admin privileges on all resources managed by Lake Formation. For more information on admin privileges, see Granting Lake Formation Permissions.

RegisterResource

Registers the resource as managed by the Data Catalog.

RemoveLFTagsFromResource

Removes a tag from the resource. Only database, table, or tableWithColumns resource are allowed. To tag columns, use the column inclusion list in tableWithColumns to specify column input.

RevokePermissions

Revokes permissions to the principal to access metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3.

SearchDatabasesByLFTags

This operation allows a search on DATABASE resources by TagCondition. This operation is used by admins who want to grant user permissions on certain TagConditions. Before making a grant, the admin can use SearchDatabasesByTags to find all resources where the given TagConditions are valid to verify whether the returned resources can be shared.

SearchTablesByLFTags

This operation allows a search on TABLE resources by LFTags. This will be used by admins who want to grant user permissions on certain LFTags. Before making a grant, the admin can use SearchTablesByLFTags to find all resources where the given LFTags are valid to verify whether the returned resources can be shared.

UpdateLFTag

Updates the list of possible values for the specified tag key. If the tag does not exist, the operation throws an EntityNotFoundException. The values in the delete key values will be deleted from list of possible values. If any value in the delete key values is attached to a resource, then API errors out with a 400 Exception - "Update not allowed". Untag the attribute before deleting the tag key's value.

UpdateResource

Updates the data access role used for vending access to the given (registered) resource in AWS Lake Formation.